The timestamp contains the nanoseconds as well. 24 bytes are required (instead of 20) and the string representation has a length of 33

+++ cmd/ksuid/main.go
@@ -128,7 +128,7 @@ func printTemplate(id ksuid.KSUID) {
 		String    string
 		Raw       string
 		Time      time.Time
-		Timestamp uint32
+		Timestamp uint64
 		Payload   string
 	}{
 		String:    id.String(),
